started refactor of pydatalog
