+++ logic.py
@@ -1,0 +1,122 @@
+from pyDatalog import pyDatalog
+import itertools
+
+TERMS = [
+    "PROPERTY",
+    "ENTITY",
+    "RELATIONSHIP",
+    "SELECT_CLAUSE",
+    "FUNCTION",
+    "DATA_SOURCE_IN_SELECT_CLAUSE",
+    "QUERY",
+    "NON_UNIQUE_PROPERTY_OF",
+    "PROPERTY_REFERENCES_ENTITY_BY",
+    "QUERY_HAS_RELATIONSHIP",
+    "RELATIONSHIP_SOURCE_ENTITY_PROPERTY",
+    "RELATIONSHIP_TARGET_ENTITY_PROPERTY",
+    "RELATIONSHIP_BETWEEN_ENTITIES",
+    "ENTITY_IN_DATA_SOURCE",
+    "SELECTION",
+    "SELECTION_HAS_LABEL",
+    "QUERY_CONTAINS_SELECTION",
+    "QUERY_HAS_SELECTION_NAME",
+    "ENTITY_IN_SELECT_CLAUSE",
+    "RELATIONSHIP_SOURCE_TARGET_PROPERTIES",
+    "REFERENCES_QUERY_NAME",
+    "PROPERTY_OF",
+    "HAS_NAME",
+    "UNIQUE",
+    "COLUMN",
+    "DATA_SOURCE",
+    "UNIQUE_PROPERTY_OF",
+    "TABLE",
+    "PROPERTY_IN_SELECT_CLAUSE",
+    "COLUMN_IN_TABLE",
+]
+
+VARIABLE_NAMES = ["X", "Y", "Z", "W", "V", "U"]
+VARIABLE_INDEXES = [str(i) for i in range(10)]
+VARIABLES = VARIABLE_NAMES + [
+    "".join(i) for i in itertools.product(VARIABLE_NAMES, VARIABLE_INDEXES)
+]
+
+pyDatalog.create_terms(",".join(TERMS + VARIABLES))
+
+QUERY_HAS_RELATIONSHIP(X0, X1) <= (
+    SELECT_CLAUSE(X0) & RELATIONSHIP(X1) & HAS_NAME(X0, X2) & (X1.query_name == X2)
+)
+
+
+RELATIONSHIP_BETWEEN_ENTITIES(X0, X1, X2) <= (
+    RELATIONSHIP(X0)
+    & ENTITY(X1)
+    & ENTITY(X2)
+    & (X0.entity_type_1 == X1)
+    & (X0.entity_type_2 == X2)
+)
+
+UNIQUE_PROPERTY_OF(X0, X1) <= (
+    PROPERTY(X0) & ENTITY(X1) & PROPERTY_OF(X0, X1) & UNIQUE(X0)
+)
+
+PROPERTY_IN_SELECT_CLAUSE(X0, X1) <= (
+    PROPERTY(X0)
+    & SELECT_CLAUSE(X1)
+    & HAS_NAME(X1, X2)
+    & REFERENCES_QUERY_NAME(X0, X2)
+)
+
+ENTITY_IN_SELECT_CLAUSE(X0, X1) <= (
+    ENTITY(X0)
+    & SELECT_CLAUSE(X1)
+    & PROPERTY_IN_SELECT_CLAUSE(X2, X1)
+    & PROPERTY_OF(X2, X0)
+)
+
+ENTITY_IN_DATA_SOURCE(X0, X1) <= (
+    ENTITY(X0)
+    & DATA_SOURCE(X1)
+    & ENTITY_IN_SELECT_CLAUSE(X0, X2)
+    & DATA_SOURCE_IN_SELECT_CLAUSE(X1, X2)
+)
+
+NON_UNIQUE_PROPERTY_OF(X0, X1) <= (
+    PROPERTY_OF(X0, X1) & ~UNIQUE_PROPERTY_OF(X0, X1)
+)
+
+SELECTION_HAS_LABEL(X0, X1) <= (SELECTION(X0) & (X1 == X0.label))
+
+RELATIONSHIP_SOURCE_ENTITY_PROPERTY(X6, X5) <= (
+    RELATIONSHIP(X6)
+    & (X0 == X6.property_name_1)
+    & SELECT_CLAUSE(X1)
+    & HAS_NAME(X1, X2)
+    & (X2 == X6.query_name)
+    & QUERY_CONTAINS_SELECTION(X1, X3)
+    & SELECTION_HAS_LABEL(X3, X0)
+    & PROPERTY(X4)
+    & UNIQUE(X4)
+    & (X4.query_name == X2)
+    & (X4.entity_selection_name == X3.label)
+    & (X4.property_name == X5)
+)
+
+RELATIONSHIP_TARGET_ENTITY_PROPERTY(X6, X5) <= (
+    RELATIONSHIP(X6)
+    & (X0 == X6.property_name_2)
+    & SELECT_CLAUSE(X1)
+    & HAS_NAME(X1, X2)
+    & (X2 == X6.query_name)
+    & QUERY_CONTAINS_SELECTION(X1, X3)
+    & SELECTION_HAS_LABEL(X3, X0)
+    & PROPERTY(X4)
+    & UNIQUE(X4)
+    & (X4.query_name == X2)
+    & (X4.entity_selection_name == X3.label)
+    & (X4.property_name == X5)
+)
+
+RELATIONSHIP_SOURCE_TARGET_PROPERTIES(X0, X1, X2) <= (
+    RELATIONSHIP_SOURCE_ENTITY_PROPERTY(X0, X1)
+    & RELATIONSHIP_TARGET_ENTITY_PROPERTY(X0, X2)
+)
